[ADD] Ouput folder to Build Advanced window, letting you give a custom path where your build will be put
[ADD] General Editor Utils Class with GetBuildPathFromConfig and FindReleaseCandidate functions
[VUP]
