+++ Assets/Editor/Newconfify.asset
@@ -1,0 +1,28 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6e8365896f232e54db8fac2eb470232c, type: 3}
+  m_Name: Newconfify
+  m_EditorClassIdentifier: 
+  Platform: 0
+  Configuration: 0
+  Scenes: []
+  OutputFormat: '{project}/{project}_{config}_{date}_{rc}'
+  OutputFolder: 
+  PreBuildSteps: []
+  PostBuildSteps: []
+  ScriptingDefineSymbols: 
+  Headless: 0
+  WaitForPlayerConnection: 0
+  IncludeTestCode: 0
+  Compression: 0
+  references:
+    version: 1
