+++ Assets/Editor/Newconfify.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69cfc17c04151384181aae78f56d501e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
