+++ Assets/Editor/OhterTestOuputConfig.asset
@@ -1,0 +1,34 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6e8365896f232e54db8fac2eb470232c, type: 3}
+  m_Name: OhterTestOuputConfig
+  m_EditorClassIdentifier: 
+  Platform: 0
+  Configuration: 2
+  Scenes:
+  - id: 0
+  OutputFormat: '{project}/{date}_{rc}/{project}_{config}'
+  OutputFolder: D:/Builds/A_Magnus_TESTING
+  PreBuildSteps: []
+  PostBuildSteps: []
+  ScriptingDefineSymbols: 
+  Headless: 0
+  WaitForPlayerConnection: 1
+  IncludeTestCode: 0
+  Compression: 0
+  references:
+    version: 1
+    00000000:
+      type: {class: SceneReferenceData, ns: Rhinox.Lightspeed, asm: com.rhinox.open.lightspeed}
+      data:
+        sceneGuid: dcc75c62e9b2c404ca564f08bd16532f
+        scenePath: Assets/Test/Scene/emptyScene.unity
