+++ Assets/Editor/OhterTestOuputConfig.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f97b43b49dbb21444b5ce473431770e8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
