+++ Assets/Editor/TestOutput.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ddc90700dcf9e14ea93e07ff31c34bf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
