+++ Assets/Magnus/Editor/Utils/MagnusUtils.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 66df1aa1d4f0487d990e79d61ca390bf
+timeCreated: 1706513430
