+++ Assets/Magnus/package.json
@@ -1,7 +1,7 @@
 {
   "name": "com.rhinox.open.magnus",
   "displayName": "Magnus - Game Framework",
-  "version": "0.8.2",
+  "version": "0.8.3",
   "unity": "2019.4",
   "description": "Game Framework built on top of Unity",
   "keywords": [
